pyproject.toml
[tool.poetry]
name = "used-car-price-prediction"
version = "0.1.0"
description = ""
authors = ["takumiw"]
readme = "README.md"
packages = [{include = "used_car_price_prediction"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
signate = "^9.9.10"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
lightgbm = "^4.0.0"
catboost = "^1.2"
loguru = "^0.7.0"
hydra-core = "^1.3.2"
python-dotenv = "^1.0.0"
torch = "^2.0.1"
torchinfo = "^1.8.0"
pytorch-lightning = "^2.0.7"
wandb = "^0.15.8"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.4"
jupyterlab-horizon-theme = "^3.1.1"
matplotlib-venn = "^0.11.9"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
flake8-fixme = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120